JseWindow: Header cleanup

+++ jse-window.h
@@ -5,11 +5,14 @@
 #include "julia.h"
 
 #define JSE_TYPE_WINDOW (jse_window_get_type ())
+
 G_DECLARE_FINAL_TYPE (JseWindow, jse_window, JSE, WINDOW, GtkApplicationWindow)
 
-JseWindow *jse_window_new       (GtkApplication *app);
-void jse_window_set_image (struct _JseWindow *win, GtkImage *img);
-void jse_window_set_zoom_level (JseWindow *win, gdouble zoom_level);
-gint jse_window_get_zoom_level (JseWindow *win);
+JseWindow *jse_window_new (GtkApplication *app);
+void jse_window_set_image (struct _JseWindow *win,
+                           GtkImage *img);
+void jse_window_set_zoom_level (JseWindow *win,
+                                gdouble zoom_level);
+gdouble jse_window_get_zoom_level (JseWindow *win);
 
 #endif /* __JSE_WINDOW_H__ */
